# docs/05-接口文档/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI 开放平台接口文档","description":"基于 OpenAPI 3.0 规范的开放平台接口文档","version":"0.0.1-SNAPSHOT"},"servers":[{"url":"http://localhost:8101","description":"Generated server url"}],"paths":{"/v1/user/update-status":{"post":{"tags":["用户管理"],"summary":"启用/禁用用户（管理员）","operationId":"updateStatus","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"enabled","in":"query","required":true,"schema":{"type":"boolean"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/v1/user/update-role":{"post":{"tags":["用户管理"],"summary":"修改用户角色（管理员）","operationId":"updateRole","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"role","in":"query","required":true,"schema":{"type":"string"},"example":"admin"}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/v1/user/register":{"post":{"tags":["用户管理"],"summary":"用户注册","operationId":"register","parameters":[{"name":"userAccount","in":"query","description":"账号","required":true,"schema":{"type":"string"},"example":"admin"},{"name":"userPassword","in":"query","description":"密码","required":true,"schema":{"type":"string"},"example":123456},{"name":"userName","in":"query","description":"昵称","required":false,"schema":{"type":"string"},"example":"管理员"}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUser"}}}}}}},"/v1/user/login":{"post":{"tags":["用户管理"],"summary":"用户登录","operationId":"login","parameters":[{"name":"userAccount","in":"query","description":"账号","required":true,"schema":{"type":"string"},"example":"admin"},{"name":"userPassword","in":"query","description":"密码","required":true,"schema":{"type":"string"},"example":123456}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMapStringObject"}}}}}}},"/v1/interface/update":{"post":{"tags":["接口管理"],"summary":"编辑接口（管理员）","operationId":"update","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"name","in":"query","required":false,"schema":{"type":"string"}},{"name":"description","in":"query","required":false,"schema":{"type":"string"}},{"name":"method","in":"query","required":false,"schema":{"type":"string"}},{"name":"url","in":"query","required":false,"schema":{"type":"string"}},{"name":"requestParams","in":"query","required":false,"schema":{"type":"string"}},{"name":"responseExample","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/v1/interface/unsubscribe":{"post":{"tags":["接口管理"],"summary":"取消订阅","operationId":"unsubscribe","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/v1/interface/subscribe":{"post":{"tags":["接口管理"],"summary":"订阅接口","operationId":"subscribe","parameters":[{"name":"interfaceId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"appId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseInterfaceSubscribe"}}}}}}},"/v1/interface/online":{"post":{"tags":["接口管理"],"summary":"接口上线（管理员）","operationId":"online","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/v1/interface/offline":{"post":{"tags":["接口管理"],"summary":"接口下线（管理员）","operationId":"offline","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/v1/interface/delete":{"post":{"tags":["接口管理"],"summary":"删除接口（管理员）","operationId":"delete","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/v1/interface/create":{"post":{"tags":["接口管理"],"summary":"新增接口（管理员）","operationId":"create","parameters":[{"name":"name","in":"query","required":true,"schema":{"type":"string"}},{"name":"description","in":"query","required":true,"schema":{"type":"string"}},{"name":"method","in":"query","required":true,"schema":{"type":"string"}},{"name":"url","in":"query","required":true,"schema":{"type":"string"}},{"name":"requestParams","in":"query","required":false,"schema":{"type":"string"}},{"name":"responseExample","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseInterfaceInfo"}}}}}}},"/v1/interface/approve":{"post":{"tags":["接口管理"],"summary":"订阅审批（管理员）","operationId":"approve","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"approved","in":"query","required":true,"schema":{"type":"boolean"},"example":true}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/v1/app/update":{"post":{"tags":["应用管理"],"summary":"修改应用名称","operationId":"update_1","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"appName","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/v1/app/update-status":{"post":{"tags":["应用管理"],"summary":"启用/禁用应用","operationId":"updateStatus_1","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"enabled","in":"query","required":true,"schema":{"type":"boolean"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/v1/app/reset-secret":{"post":{"tags":["应用管理"],"summary":"重置 SecretKey","operationId":"resetSecret","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseApp"}}}}}}},"/v1/app/delete":{"post":{"tags":["应用管理"],"summary":"删除应用","operationId":"delete_1","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/v1/app/create":{"post":{"tags":["应用管理"],"summary":"创建应用（生成密钥）","operationId":"create_1","parameters":[{"name":"appName","in":"query","description":"应用名称","required":true,"schema":{"type":"string"},"example":"我的应用"},{"name":"userId","in":"query","description":"用户 ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseApp"}}}}}}},"/api/demo/echo":{"post":{"tags":["演示接口"],"summary":"参数回显","operationId":"echo","parameters":[{"name":"params","in":"query","description":"任意表单参数","required":true,"schema":{"type":"object","additionalProperties":{"type":"string"}},"example":"hello=world"}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMapStringString"}}}}}}},"/v1/user/list":{"get":{"tags":["用户管理"],"summary":"用户列表（管理员）","operationId":"listUsers","parameters":[{"name":"keyword","in":"query","description":"搜索关键字（账号/昵称）","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListUser"}}}}}}},"/v1/stats/overview":{"get":{"tags":["调用统计"],"summary":"调用统计概览","operationId":"overview","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMapStringObject"}}}}}}},"/v1/stats/daily":{"get":{"tags":["调用统计"],"summary":"近 N 天调用趋势","operationId":"daily","parameters":[{"name":"days","in":"query","description":"天数","required":false,"schema":{"type":"integer","format":"int32","default":7},"example":7}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListMapStringObject"}}}}}}},"/v1/interface/{id}":{"get":{"tags":["接口管理"],"summary":"查询接口详情","operationId":"detail","parameters":[{"name":"id","in":"path","description":"接口 ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseInterfaceInfo"}}}}}}},"/v1/interface/subscribes":{"get":{"tags":["接口管理"],"summary":"订阅审批列表（管理员）","operationId":"subscribes","parameters":[{"name":"status","in":"query","description":"状态：0待审批 1通过 2拒绝","required":false,"schema":{"type":"integer","format":"int32"},"example":0}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListMapStringObject"}}}}}}},"/v1/interface/my-subscribes":{"get":{"tags":["接口管理"],"summary":"我的订阅列表","operationId":"mySubscribes","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListMapStringObject"}}}}}}},"/v1/interface/list":{"get":{"tags":["接口管理"],"summary":"查询已上线接口","operationId":"listOnline","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListInterfaceInfo"}}}}}}},"/v1/interface/list-all":{"get":{"tags":["接口管理"],"summary":"全部接口列表（管理员）","operationId":"listAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListInterfaceInfo"}}}}}}},"/v1/app/list":{"get":{"tags":["应用管理"],"summary":"查询用户的应用列表","operationId":"list","parameters":[{"name":"userId","in":"query","description":"用户 ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListApp"}}}}}}},"/api/demo/name":{"get":{"tags":["演示接口"],"summary":"随机名称","operationId":"randomName","parameters":[{"name":"prefix","in":"query","description":"名称前缀","required":false,"schema":{"type":"string"},"example":"Hi-"}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseString"}}}}}}}},"components":{"schemas":{"ApiResponseVoid":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"object"}}},"ApiResponseUser":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/User"}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"userAccount":{"type":"string"},"userPassword":{"type":"string"},"userName":{"type":"string"},"userRole":{"type":"string"},"status":{"type":"integer","format":"int32"},"createTime":{"type":"string","format":"date-time"},"updateTime":{"type":"string","format":"date-time"},"isDelete":{"type":"integer","format":"int32"}}},"ApiResponseMapStringObject":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"object","additionalProperties":{"type":"object"}}}},"ApiResponseInterfaceSubscribe":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/InterfaceSubscribe"}}},"InterfaceSubscribe":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"interfaceId":{"type":"integer","format":"int64"},"appId":{"type":"integer","format":"int64"},"userId":{"type":"integer","format":"int64"},"status":{"type":"integer","format":"int32"},"createTime":{"type":"string","format":"date-time"},"updateTime":{"type":"string","format":"date-time"},"isDelete":{"type":"integer","format":"int32"}}},"ApiResponseInterfaceInfo":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/InterfaceInfo"}}},"InterfaceInfo":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"method":{"type":"string"},"url":{"type":"string"},"requestParams":{"type":"string"},"responseExample":{"type":"string"},"status":{"type":"integer","format":"int32"},"createTime":{"type":"string","format":"date-time"},"updateTime":{"type":"string","format":"date-time"},"isDelete":{"type":"integer","format":"int32"}}},"ApiResponseApp":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/App"}}},"App":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"appName":{"type":"string"},"accessKey":{"type":"string"},"secretKey":{"type":"string"},"userId":{"type":"integer","format":"int64"},"status":{"type":"integer","format":"int32"},"createTime":{"type":"string","format":"date-time"},"updateTime":{"type":"string","format":"date-time"},"isDelete":{"type":"integer","format":"int32"}}},"ApiResponseMapStringString":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"object","additionalProperties":{"type":"string"}}}},"ApiResponseListUser":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}},"ApiResponseListMapStringObject":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"array","items":{"type":"object","additionalProperties":{"type":"object"}}}}},"ApiResponseListInterfaceInfo":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/InterfaceInfo"}}}},"ApiResponseListApp":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/App"}}}},"ApiResponseString":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"string"}}}}}}